Added scheduler, plotted losses, loading model and new layers.
Scheduler was added to the training process as well as printing out validation loss and saving both val and train loss so that they can be visualized after traingin. The model was saved and loading was added and checked and it got a 88.3% accuracy but on the validation data. Added an additional dropout, batchnorm and linear layer.